More fixes for INA and Arduino battery status

+++ test_arduino.py
@@ -15,13 +15,14 @@ def read():
     ser = serial.Serial('/dev/ttyACM0', 9600)
     values = []
     for i in range(1, 15):
-     values.append(float(ser.readline().strip(",\n\r")))
+     values.append(int(ser.readline().strip(",\n\r")))
     ser.close()
-    return float(sum(values)) / max(len(values), 1)
+    return int(sum(values)) / max(len(values), 1)
     exit()
 
 while True:
 	ret = round(read() + 10)
 	print ret
 	sleep(REFRESH_RATE)
+
